Get rid of old fill_triangle routine and make unit tests pass

+++ librend/rend.h
@@ -24,9 +24,9 @@ public:
                    double x1, double y1,
                    const Color& color);
     void fill_triangle(ImageInterface &img,
-                   int x0, int y0, 
-                   int x1, int y1,
-                   int x2, int y2,
+                   double x0, double y0, 
+                   double x1, double y1,
+                   double x2, double y2,
                    const Color& color);
     void fill_triangle_barycentric(ImageInterface &img,
                    double x0, double y0,
